Move registration functionality to separate files
