+++ simulator/tests/auth.test.js
@@ -7,10 +7,6 @@ const sinonChai = require("sinon-chai");
 const jwt = require("jsonwebtoken");
 
 const {
-  registerProsumer,
-  registerManager,
-  prosumerRegistrationAccountQuery,
-  managerRegistrationAccountsQuery,
   logInUser,
   authenticateLoggedIn,
   authenticateIsMe,
@@ -218,148 +214,6 @@ describe("auth", () => {
           const token = await mockModule.logInUser("email", "password");
           const user = jwt.verify(token, "secret");
           expect(user.accountId).to.be.equal(8);
-          expect(user.manager).to.be.false;
-        });
-      });
-    });
-  });
-
-  describe("registration", () => {
-    describe("manager", () => {
-      describe("unsuccessful", () => {
-        afterEach(() => {
-          sinon.restore();
-        });
-
-        it("should return null if email password is passed as null", async () => {
-          expect(await registerManager(null, "password")).to.be.equal(null);
-        });
-
-        it("should return null if no password is passed", async () => {
-          expect(await registerManager("email")).to.be.equal(null);
-        });
-
-        it("should return null if no manager password is passed", async () => {
-          expect(
-            await registerManager("email", "password", "managerPassword")
-          ).to.be.equal(null);
-        });
-
-        it("should return null if an error is raised", async () => {
-          const mockClient = {
-            query(queryString) {
-              // Don't throw in catch clause
-              if (queryString !== "ROLLBACK") {
-                throw new Error("test");
-              }
-            }
-          };
-
-          sinon.stub(pool, "connect").resolves(mockClient);
-          expect(await registerManager()).to.be.equal(null);
-        });
-      });
-
-      describe("successful", () => {
-        afterEach(() => {
-          sinon.restore();
-        });
-
-        it("should return a token containing an account id and manager flag set to true when succesful", async () => {
-          const authModule = rewire("../src/auth");
-          authModule.__set__({
-            bcrypt: {
-              genSaltSync: () => "salt",
-              hashSync: () => "testHash"
-            },
-            // Make sure the private key and manager password is not
-            // set via environment
-            process: { env: {} }
-          });
-
-          const mockClient = {
-            async query(queryString) {
-              if (queryString === managerRegistrationAccountsQuery) {
-                return { rows: [{ id: 7 }] };
-              }
-              return null;
-            },
-            release() {}
-          };
-
-          sinon.stub(pool, "connect").resolves(mockClient);
-
-          const token = await authModule.registerManager(
-            "email",
-            "password",
-            "manager"
-          );
-          const user = jwt.verify(token, "secret");
-          expect(user.accountId).to.be.equal(7);
-          expect(user.manager).to.be.true;
-        });
-      });
-    });
-
-    describe("prosumer", () => {
-      describe("unsuccessful", () => {
-        afterEach(() => {
-          sinon.restore();
-        });
-
-        it("should return null if email password is passed as null", async () => {
-          expect(await registerProsumer(null, "password")).to.be.equal(null);
-        });
-
-        it("should return null if no password is passed", async () => {
-          expect(await registerProsumer("email")).to.be.equal(null);
-        });
-
-        it("should return null if an error is raised", async () => {
-          const mockClient = {
-            query(queryString) {
-              // Don't throw in catch clause
-              if (queryString !== "ROLLBACK") {
-                throw new Error("test");
-              }
-            }
-          };
-
-          sinon.stub(pool, "connect").resolves(mockClient);
-          expect(await registerProsumer()).to.be.equal(null);
-        });
-      });
-
-      describe("successful", () => {
-        afterEach(() => {
-          sinon.restore();
-        });
-
-        it("should return a token containing an account id and manager flag set to false when succesful", async () => {
-          const authModule = rewire("../src/auth");
-          authModule.__set__({
-            bcrypt: {
-              genSaltSync: () => "salt",
-              hashSync: () => "testHash"
-            },
-            process: { env: {} } // Make sure the private key is not set via environment
-          });
-
-          const mockClient = {
-            async query(queryString) {
-              if (queryString === prosumerRegistrationAccountQuery) {
-                return { rows: [{ id: 7 }] };
-              }
-              return null;
-            },
-            release() {}
-          };
-
-          sinon.stub(pool, "connect").resolves(mockClient);
-
-          const token = await authModule.registerProsumer("email", "password");
-          const user = jwt.verify(token, "secret");
-          expect(user.accountId).to.be.equal(7);
           expect(user.manager).to.be.false;
         });
       });
